requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.49.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
environ==1.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.1
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.8
